html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-form-submit.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: form submit</title>
<iframe name="iframe" src="support/blank.htm?src" loading="lazy"></iframe>
<form action="support/blank.htm" target="iframe"></form>
<script>
const iframe = document.querySelector('iframe');
const form = document.querySelector('form');
form.submit();
</script>
<!-- Loading testharness.js here is intentional. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?", location.href).href);
    done();
  }, 250);
};
</script>
